+++ CHANGES.rst
@@ -12,6 +12,7 @@ Improvements:
  - Add option to choose default media source (#2763)
  - Add option to choose default photo compression (#2763)
  - Add option to disable camera shutter sound
+ - Auto-refresh scalar token when a 403 error is detected (#3051)
 
 Other changes:
  -